refator playground context
